Clean up importModule: extract runtime, deduplicate, cargo fmt
- Extract the mini Turbopack runtime from webpack-loaders.ts into
  webpack-loaders-runtime.ts with a single executeModules() entry point
- Deduplicate module ID regex into MODULE_ID_PATH_RE constant and
  resolveModuleIdPath() helper
- Extract ensureModule() to eliminate boilerplate in ctx.s/ctx.v/ctx.n
- Remove redundant require('fs') (already imported at top level)
- Run cargo fmt on webpack.rs
