+++ turbopack/crates/turbopack-node/js/src/transforms/webpack-loaders.ts
@@ -15,6 +15,10 @@ import {
   toPath,
   type TransformIpc,
 } from './transforms'
+import {
+  executeModules,
+  type ImportModuleResult,
+} from './webpack-loaders-runtime'
 import fs from 'fs'
 import path from 'path'
 
@@ -388,468 +392,9 @@ const transform = (
                 type: 'importModule',
                 lookupPath: toPath(resourceDir),
                 request: actualRequest,
-              })) as {
-                entryId: string
-                modules: Array<{
-                  id: string
-                  code: string
-                  sourceMap?: string
-                  moduleAndExports: boolean
-                }>
-              }
+              })) as ImportModuleResult
 
-              // Build a mini Turbopack runtime to execute the compiled
-              // module code, similar to webpack's executeModule.
-              const vm = require('vm')
-              const nodeUrl = require('url')
-
-              type ModuleObj = {
-                exports: any
-                error: any
-                id: string
-                namespaceObject: any
-              }
-
-              // Module cache and factory map
-              const moduleCache = new Map<string, ModuleObj>()
-              const moduleFactories = new Map<
-                string,
-                { factory: Function; moduleAndExports: boolean }
-              >()
-
-              // Map from exported asset URLs to the module IDs that
-              // produced them (populated by ctx.q)
-              const urlToModuleId = new Map<string, string>()
-
-              // Resolve a [project]-relative path to an absolute
-              // filesystem path. The module ID contains a path
-              // relative to Turbopack's project root which may be
-              // an ancestor of contextDir (process.cwd()).
-              function resolveProjectPath(
-                relPath: string
-              ): string {
-                const fs = require('fs')
-                let candidate = path.resolve(
-                  contextDir,
-                  relPath
-                )
-                if (fs.existsSync(candidate)) return candidate
-                let dir = contextDir
-                while (true) {
-                  const parent = path.dirname(dir)
-                  if (parent === dir) break
-                  candidate = path.resolve(parent, relPath)
-                  if (fs.existsSync(candidate))
-                    return candidate
-                  dir = parent
-                }
-                throw new Error(
-                  `importModule: cannot resolve project path: ${relPath}`
-                )
-              }
-
-              // Compile all module factories
-              for (const mod of result.modules) {
-                const factoryCode = mod.moduleAndExports
-                  ? `(function(__turbopack_context__, module, exports) {\n${mod.code}\n})`
-                  : `(function(__turbopack_context__) {\n${mod.code}\n})`
-                const factory = vm.runInThisContext(factoryCode, {
-                  filename: mod.id,
-                })
-                moduleFactories.set(mod.id, {
-                  factory,
-                  moduleAndExports: mod.moduleAndExports,
-                })
-              }
-
-              // ESM helpers
-              function defineProp(
-                obj: any,
-                name: string,
-                options: PropertyDescriptor
-              ) {
-                if (!Object.prototype.hasOwnProperty.call(obj, name)) {
-                  Object.defineProperty(obj, name, options)
-                }
-              }
-
-              function esmBindings(
-                exports: any,
-                bindings: any[]
-              ) {
-                defineProp(exports, '__esModule', { value: true })
-                let i = 0
-                while (i < bindings.length) {
-                  const propName = bindings[i++]
-                  const tagOrFunction = bindings[i++]
-                  if (typeof tagOrFunction === 'number') {
-                    if (tagOrFunction === 0) {
-                      defineProp(exports, propName, {
-                        value: bindings[i++],
-                        enumerable: true,
-                        writable: false,
-                      })
-                    }
-                  } else {
-                    const getter = tagOrFunction
-                    if (
-                      i < bindings.length &&
-                      typeof bindings[i] === 'function'
-                    ) {
-                      const setter = bindings[i++]
-                      defineProp(exports, propName, {
-                        get: getter,
-                        set: setter,
-                        enumerable: true,
-                      })
-                    } else {
-                      defineProp(exports, propName, {
-                        get: getter,
-                        enumerable: true,
-                      })
-                    }
-                  }
-                }
-              }
-
-              function interopEsm(
-                raw: any,
-                allowExportDefault?: boolean
-              ): any {
-                const ns = Object.create(null)
-                defineProp(ns, '__esModule', { value: true })
-                if (
-                  raw &&
-                  (typeof raw === 'object' || typeof raw === 'function')
-                ) {
-                  for (const key of Object.getOwnPropertyNames(raw)) {
-                    defineProp(ns, key, {
-                      enumerable: true,
-                      get: () => raw[key],
-                    })
-                  }
-                }
-                if (
-                  !(
-                    allowExportDefault &&
-                    raw &&
-                    Object.prototype.hasOwnProperty.call(raw, 'default')
-                  )
-                ) {
-                  defineProp(ns, 'default', {
-                    value: raw,
-                    enumerable: true,
-                  })
-                }
-                return ns
-              }
-
-              // Symbol used to track async module promises on
-              // namespace/export objects so handleAsyncDependencies
-              // can detect and await them.
-              const ASYNC_PROMISE = Symbol('asyncPromise')
-
-              // Instantiate a module by ID
-              function instantiateModule(id: string): ModuleObj {
-                const cached = moduleCache.get(id)
-                if (cached) return cached
-
-                const entry = moduleFactories.get(id)
-                if (!entry) {
-                  throw new Error(
-                    `importModule: module not found: ${id}`
-                  )
-                }
-
-                const moduleObj: ModuleObj = {
-                  exports: {},
-                  error: undefined,
-                  id,
-                  namespaceObject: undefined,
-                }
-                moduleCache.set(id, moduleObj)
-
-                // Create __turbopack_context__
-                const ctx: any = Object.create(null)
-                ctx.m = moduleObj
-                ctx.e = moduleObj.exports
-                ctx.c = moduleCache
-                ctx.M = moduleFactories
-                ctx.g = globalThis
-
-                // CommonJS require
-                ctx.r = (depId: string) => {
-                  return instantiateModule(depId).exports
-                }
-
-                // ESM import
-                ctx.i = (depId: string) => {
-                  const mod = instantiateModule(depId)
-                  if (mod.namespaceObject)
-                    return mod.namespaceObject
-                  const raw = mod.exports
-                  return (mod.namespaceObject = interopEsm(
-                    raw,
-                    raw && raw.__esModule
-                  ))
-                }
-
-                // ESM export bindings
-                ctx.s = (bindings: any[], targetId?: string) => {
-                  let targetModule: ModuleObj
-                  let targetExports: any
-                  if (targetId != null) {
-                    targetModule = moduleCache.get(targetId) || {
-                      exports: {},
-                      error: undefined,
-                      id: targetId,
-                      namespaceObject: undefined,
-                    }
-                    if (!moduleCache.has(targetId)) {
-                      moduleCache.set(targetId, targetModule)
-                    }
-                    targetExports = targetModule.exports
-                  } else {
-                    targetModule = moduleObj
-                    targetExports = moduleObj.exports
-                  }
-                  targetModule.namespaceObject = targetExports
-                  esmBindings(targetExports, bindings)
-                }
-
-                // Export value
-                ctx.v = (value: any, targetId?: string) => {
-                  if (targetId != null) {
-                    const mod = moduleCache.get(targetId) || {
-                      exports: {},
-                      error: undefined,
-                      id: targetId,
-                      namespaceObject: undefined,
-                    }
-                    mod.exports = value
-                    if (!moduleCache.has(targetId)) {
-                      moduleCache.set(targetId, mod)
-                    }
-                  } else {
-                    moduleObj.exports = value
-                  }
-                }
-
-                // Export namespace
-                ctx.n = (namespace: any, targetId?: string) => {
-                  if (targetId != null) {
-                    const mod = moduleCache.get(targetId) || {
-                      exports: {},
-                      error: undefined,
-                      id: targetId,
-                      namespaceObject: undefined,
-                    }
-                    mod.exports = mod.namespaceObject = namespace
-                    if (!moduleCache.has(targetId)) {
-                      moduleCache.set(targetId, mod)
-                    }
-                  } else {
-                    moduleObj.exports = moduleObj.namespaceObject =
-                      namespace
-                  }
-                }
-
-                // Dynamic export
-                ctx.j = (object: any) => {
-                  if (
-                    typeof object === 'object' &&
-                    object !== null
-                  ) {
-                    for (const key of Object.keys(object)) {
-                      if (key !== 'default') {
-                        defineProp(moduleObj.exports, key, {
-                          enumerable: true,
-                          get: () => object[key],
-                        })
-                      }
-                    }
-                  }
-                }
-
-                // External require (use __turbopack_external_require__
-                // to avoid Turbopack static analysis of require())
-                ctx.x = __turbopack_external_require__
-
-                // Runtime require
-                ctx.t = __turbopack_external_require__
-
-                // Require stub (throws in ESM)
-                ctx.z = () => {
-                  throw new Error(
-                    'dynamic usage of require is not supported'
-                  )
-                }
-
-                // Export URL (for static assets and raw wasm modules)
-                ctx.q = (url: string) => {
-                  defineProp(moduleObj.exports, 'default', {
-                    value: url,
-                    enumerable: true,
-                  })
-                  urlToModuleId.set(url, moduleObj.id)
-                }
-
-                // Resolve module ID path (for new URL(./file,
-                // import.meta.url) patterns). Returns a file:// URL
-                // pointing to the original source file.
-                ctx.R = (depModuleId: string) => {
-                  const exported = ctx.r(depModuleId)
-                  const assetUrl =
-                    exported?.default ?? exported
-                  const sourceModId =
-                    urlToModuleId.get(assetUrl)
-                  if (sourceModId) {
-                    const m = sourceModId.match(
-                      /^\[project\]\/(.+?)(?:\s+\[|\s+\()/
-                    )
-                    if (m) {
-                      const absPath = resolveProjectPath(
-                        m[1]
-                      )
-                      return nodeUrl.pathToFileURL(absPath)
-                        .href
-                    }
-                  }
-                  return assetUrl
-                }
-
-                // Load and instantiate WebAssembly asynchronously.
-                // The chunkPath comes from a raw wasm module's
-                // ctx.q() export.
-                ctx.w = async (
-                  chunkPath: string,
-                  _edgeModule: any,
-                  importsObj: any
-                ) => {
-                  const sourceModId =
-                    urlToModuleId.get(chunkPath)
-                  if (!sourceModId) {
-                    throw new Error(
-                      `importModule: wasm source not found for ${chunkPath}`
-                    )
-                  }
-                  const m = sourceModId.match(
-                    /^\[project\]\/(.+?)(?:\s+\[|\s+\()/
-                  )
-                  if (!m) {
-                    throw new Error(
-                      `importModule: cannot extract path from module ID: ${sourceModId}`
-                    )
-                  }
-                  const absPath = resolveProjectPath(m[1])
-                  const wasmBuffer = require('fs').readFileSync(
-                    absPath
-                  )
-                  const { instance } =
-                    await WebAssembly.instantiate(
-                      wasmBuffer,
-                      importsObj || {}
-                    )
-                  return instance.exports
-                }
-
-                // Async module handler. When a module has top-level
-                // await or imports async dependencies, the code
-                // generator wraps its body with ctx.a(). This sets
-                // up a promise that tracks the module's async
-                // initialization and replaces module.exports with a
-                // proxy object tagged with the promise so that
-                // handleAsyncDependencies can detect and await it.
-                ctx.a = (
-                  body: Function,
-                  hasAwait: boolean
-                ) => {
-                  let resolvePromise!: () => void
-                  let rejectPromise!: (err: any) => void
-                  const asyncPromise = new Promise<void>(
-                    (resolve, reject) => {
-                      resolvePromise = resolve
-                      rejectPromise = reject
-                    }
-                  )
-
-                  // Replace module exports with a proxy object
-                  // tagged with the async promise. ESM bindings
-                  // (ctx.s) will define getters on this proxy,
-                  // and handleAsyncDependencies will detect the
-                  // ASYNC_PROMISE symbol to await completion.
-                  const exportProxy = Object.create(null)
-                  ;(exportProxy as any)[ASYNC_PROMISE] =
-                    asyncPromise
-                  moduleObj.exports = exportProxy
-                  moduleObj.namespaceObject = exportProxy
-
-                  function handleAsyncDependencies(
-                    deps: any[]
-                  ) {
-                    const promises: Promise<void>[] = []
-                    for (const dep of deps) {
-                      const p =
-                        dep && dep[ASYNC_PROMISE]
-                      if (p) promises.push(p)
-                    }
-                    if (promises.length > 0) {
-                      return Promise.all(promises).then(
-                        () => () => deps
-                      )
-                    }
-                    return deps
-                  }
-
-                  function asyncResult(err?: any) {
-                    if (err) rejectPromise(err)
-                    else resolvePromise()
-                  }
-
-                  body(handleAsyncDependencies, asyncResult)
-                }
-
-                // Dynamic import
-                ctx.l = (depId: string) => {
-                  const mod = instantiateModule(depId)
-                  const ns =
-                    mod.namespaceObject ||
-                    (mod.namespaceObject = interopEsm(
-                      mod.exports,
-                      mod.exports && mod.exports.__esModule
-                    ))
-                  const p = ns && ns[ASYNC_PROMISE]
-                  if (p) return p.then(() => ns)
-                  return Promise.resolve(ns)
-                }
-
-                // Execute the factory
-                if (entry.moduleAndExports) {
-                  entry.factory(
-                    ctx,
-                    moduleObj,
-                    moduleObj.exports
-                  )
-                } else {
-                  entry.factory(ctx)
-                }
-
-                return moduleObj
-              }
-
-              // Execute the entry module and return its exports
-              const entryModule = instantiateModule(
-                result.entryId
-              )
-              // If the module is async, wait for it to complete
-              const ns =
-                entryModule.namespaceObject ||
-                entryModule.exports
-              const asyncP = ns && ns[ASYNC_PROMISE]
-              if (asyncP) {
-                await asyncP
-              }
-              return ns
+              return executeModules(result, contextDir)
             }
 
             if (!callback) {
